[FEATURE] 5.3.0 support

+++ UnsafeCommandPack_module/src/main/java/pro/gravit/launchermodules/unsafecommands/commands/LwjglDownloadCommand.java
@@ -7,6 +7,7 @@ import pro.gravit.launchserver.LaunchServer;
 import pro.gravit.launchserver.command.Command;
 import pro.gravit.utils.Downloader;
 import pro.gravit.utils.helper.IOHelper;
+import pro.gravit.utils.helper.JVMHelper;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -38,6 +39,43 @@ public class LwjglDownloadCommand extends Command {
         return "download lwjgl 3.3.0+";
     }
 
+    public record OSArchPair(JVMHelper.ARCH arch, JVMHelper.OS os) {
+
+    }
+
+    public static OSArchPair getFromLwjglNativeName(String name) {
+        int pos = name.indexOf("-");
+        JVMHelper.ARCH arch;
+        JVMHelper.OS os;
+        if(pos < 0) {
+            arch = JVMHelper.ARCH.X86_64;
+            os = getOS(name);
+        } else {
+            os = getOS(name.substring(0, pos));
+            arch = getArch(name.substring(pos+1));
+        }
+        return new OSArchPair(arch, os);
+    }
+
+    private static JVMHelper.ARCH getArch(String name) {
+        return switch (name) {
+            case "x86" -> JVMHelper.ARCH.X86;
+            case "arm64" -> JVMHelper.ARCH.ARM64;
+            case "arm32" -> JVMHelper.ARCH.ARM32;
+            case "x86-64" -> JVMHelper.ARCH.X86_64;
+            default -> throw new IllegalArgumentException(name);
+        };
+    }
+
+    private static JVMHelper.OS getOS(String name) {
+        return switch (name) {
+            case "windows" -> JVMHelper.OS.MUSTDIE;
+            case "linux" -> JVMHelper.OS.LINUX;
+            case "macos" -> JVMHelper.OS.MACOSX;
+            default -> throw new IllegalArgumentException(name);
+        };
+    }
+
     @Override
     public void invoke(String... args) throws Exception {
         verifyArgs(args, 2);
@@ -46,7 +84,7 @@ public class LwjglDownloadCommand extends Command {
         Path lwjglDir = clientDir.resolve("libraries").resolve("org").resolve("lwjgl");
         Path natives = clientDir.resolve("natives");
         List<String> components = List.of("lwjgl", "lwjgl-stb", "lwjgl-opengl", "lwjgl-openal", "lwjgl-glfw", "lwjgl-tinyfd", "lwjgl-jemalloc");
-        List<String> archs = List.of("linux", "windows", "windows-x86", "macos");
+        List<String> archs = List.of("linux", "windows", "windows-x86", "windows-arm64", "macos", "macos-arm64", "linux-arm64", "linux-arm32");
         String mirror = "https://repo1.maven.org/maven2/org/lwjgl/";
         for(String component : components) {
             Path jarPath = lwjglDir.resolve(component).resolve(version).resolve(component.concat("-").concat(version).concat(".jar"));
@@ -61,15 +99,19 @@ public class LwjglDownloadCommand extends Command {
                     .concat(String.format("%s-%s.jar", component, version)));
             logger.info("Download {} to {}", jarUrl, jarPath);
             download(jarUrl, jarPath);
-            List<String> processedFiles = new ArrayList<>();
             for(String arch : archs) {
                 URL nativesUrl = new URL(prepareUrl
                         .concat(String.format("%s-%s-natives-%s.jar", component, version, arch)));
+                var pair = getFromLwjglNativeName(arch);
+                Path nativesPath = natives.resolve(pair.os.name.toLowerCase()).resolve(pair.arch.name.toLowerCase());
+                IOHelper.createParentDirs(nativesPath);
                 logger.info("Download natives {}", nativesUrl);
+                List<String> processedFiles = new ArrayList<>();
                 try(ZipInputStream input = new ZipInputStream(IOHelper.newInput(nativesUrl))) {
                     ZipEntry entry = input.getNextEntry();
                     while(entry != null) {
-                        if(!entry.isDirectory() && !entry.getName().startsWith("META-INF")) {
+                        if(!entry.isDirectory() && !entry.getName().startsWith("META-INF")
+                                && !entry.getName().endsWith(".sha1") && !entry.getName().endsWith(".git")) {
                             Path path = Paths.get(entry.getName());
                             String filename = path.getFileName().toString();
                             logger.info("Process {}", filename);
@@ -83,7 +125,7 @@ public class LwjglDownloadCommand extends Command {
                                     logger.warn("Duplicate {}", filename);
                                 }
                             }
-                            IOHelper.transfer(input, natives.resolve(filename));
+                            IOHelper.transfer(input, nativesPath.resolve(filename));
                             processedFiles.add(filename);
                         }
                         entry = input.getNextEntry();
@@ -95,6 +137,7 @@ public class LwjglDownloadCommand extends Command {
     }
 
     private void download(URL url, Path target) throws IOException {
+        if(Files.exists(target)) return;
         try(InputStream input = IOHelper.newInput(url)) {
             try(OutputStream output = new FileOutputStream(target.toFile())) {
                 IOHelper.transfer(input, output);
